# [OPTIMIZATION] Optimizes the single_query_cached_kv_attention kernel